+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [4.18.0] - 2021-03-21
 ### `jsonschema-generator`
 #### Changed
 - Increase of Jackson dependency version to 2.12.1
@@ -477,7 +477,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Indicate a number's "exclusiveMaximum" according to `@DecimalMax` or `@Negative`
 
 
-[Unreleased]: https://github.com/victools/jsonschema-generator/compare/v4.17.0...HEAD
+[4.18.0]: https://github.com/victools/jsonschema-generator/compare/v4.17.0...v4.18.0
 [4.17.0]: https://github.com/victools/jsonschema-generator/compare/v4.16.0...v4.17.0
 [4.16.0]: https://github.com/victools/jsonschema-generator/compare/v4.15.1...v4.16.0
 [4.15.1]: https://github.com/victools/jsonschema-generator/compare/v4.15.0...v4.15.1